+++ src/widgets/TbDropdown.php
@@ -65,12 +65,28 @@ class TbDropdown extends TbBaseMenu {
 			$item['url'] = '#';
 		} */
 
-		$item['linkOptions']['tabindex'] = -1;
+		// Bootstrap 3's dropdown markup put tabindex="-1" on every item; Bootstrap 5 does not,
+		// and keeping it would take the whole menu out of the keyboard tab order.
+
+		// This class overrides renderMenuItem() wholesale, so the anchor styling the parent
+		// applies has to be repeated here.
+		self::addCssClass($item['linkOptions'], $this->getLinkCssClass());
+
+		if (!empty($item['active']) && $this->activeCssClass != '') {
+			self::addCssClass($item['linkOptions'], $this->activeCssClass);
+			$item['linkOptions']['aria-current'] = 'true';
+		}
+
+		if (isset($item['disabled'])) {
+			self::addCssClass($item['linkOptions'], 'disabled');
+			$item['linkOptions']['tabindex'] = '-1';
+			$item['linkOptions']['aria-disabled'] = 'true';
+		}
 
 		if (isset($item['url'])) {
 			return CHtml::link($item['label'], $item['url'], $item['linkOptions']);
 		} else {
-			return CHtml::link($item['label'], '#', array());
+			return CHtml::link($item['label'], '#', $item['linkOptions']);
 		}
 	}
 
@@ -82,7 +98,27 @@ class TbDropdown extends TbBaseMenu {
 	 */
 	public function getDividerCssClass()
 	{
-		return 'divider';
+		return 'dropdown-divider';
+	}
+
+	/**
+	 * Dropdown items take no class of their own; Bootstrap styles the anchor.
+	 *
+	 * @return string
+	 * @since 5.0.0
+	 */
+	public function getItemCssClass()
+	{
+		return '';
+	}
+
+	/**
+	 * @return string
+	 * @since 5.0.0
+	 */
+	public function getLinkCssClass()
+	{
+		return 'dropdown-item';
 	}
 
 	/**